Fix <M-w> with spaces behind cursor and \<CR>
<M-w> should delete until the end of the word before the cursor if only
whitespace is between the cursor and the previous word.

Typing \<CR> when <Leader> is mapped in insert mode makes Vim think the
insert slash is whitespace because it waits to move the cursor until
after it knows if the mapping was used. Mapping \<CR> to \<CR> is ugly
but fixes the problem.